--- test/jb_erc20_payment_terminal/set_fee_gauge.test.js
@@ -1,76 +1,0 @@
-import { expect } from 'chai';
-import { ethers } from 'hardhat';
-
-import { deployMockContract } from '@ethereum-waffle/mock-contract';
-
-import jbDirectory from '../../artifacts/contracts/interfaces/IJBDirectory.sol/IJBDirectory.json';
-import JBPaymentTerminalStore from '../../artifacts/contracts/JBPaymentTerminalStore.sol/JBPaymentTerminalStore.json';
-import jbFeeGauge from '../../artifacts/contracts/interfaces/IJBFeeGauge.sol/IJBFeeGauge.json';
-import jbOperatoreStore from '../../artifacts/contracts/interfaces/IJBOperatorStore.sol/IJBOperatorStore.json';
-import jbProjects from '../../artifacts/contracts/interfaces/IJBProjects.sol/IJBProjects.json';
-import jbSplitsStore from '../../artifacts/contracts/interfaces/IJBSplitsStore.sol/IJBSplitsStore.json';
-
-describe('JBERC20PaymentTerminal::setFeeGauge(...)', function () {
-  async function setup() {
-    let [deployer, terminalOwner, caller] = await ethers.getSigners();
-
-    let [
-      mockJbDirectory,
-      mockJBPaymentTerminalStore,
-      mockJbFeeGauge,
-      mockJbOperatorStore,
-      mockJbProjects,
-      mockJbSplitsStore,
-    ] = await Promise.all([
-      deployMockContract(deployer, jbDirectory.abi),
-      deployMockContract(deployer, JBPaymentTerminalStore.abi),
-      deployMockContract(deployer, jbFeeGauge.abi),
-      deployMockContract(deployer, jbOperatoreStore.abi),
-      deployMockContract(deployer, jbProjects.abi),
-      deployMockContract(deployer, jbSplitsStore.abi),
-    ]);
-
-    let jbTerminalFactory = await ethers.getContractFactory('JBERC20PaymentTerminal', deployer);
-
-    const currentNonce = await ethers.provider.getTransactionCount(deployer.address);
-    const futureTerminalAddress = ethers.utils.getContractAddress({
-      from: deployer.address,
-      nonce: currentNonce + 1,
-    });
-
-    await mockJBPaymentTerminalStore.mock.claimFor.withArgs(futureTerminalAddress).returns();
-
-    let jbERC20PaymentTerminal = await jbTerminalFactory
-      .connect(deployer)
-      .deploy(
-        mockJbOperatorStore.address,
-        mockJbProjects.address,
-        mockJbDirectory.address,
-        mockJbSplitsStore.address,
-        mockJBPaymentTerminalStore.address,
-        terminalOwner.address,
-      );
-
-    return {
-      terminalOwner,
-      caller,
-      jbERC20PaymentTerminal,
-      mockJbFeeGauge,
-    };
-  }
-
-  it('Should set the fee gauge and emit event if caller is terminal owner', async function () {
-    const { terminalOwner, jbERC20PaymentTerminal, mockJbFeeGauge } = await setup();
-
-    expect(await jbERC20PaymentTerminal.connect(terminalOwner).setFeeGauge(mockJbFeeGauge.address))
-      .to.emit(jbERC20PaymentTerminal, 'SetFeeGauge')
-      .withArgs(mockJbFeeGauge.address, terminalOwner.address);
-  });
-  it("Can't set the fee gauge if caller is not the terminal owner", async function () {
-    const { caller, jbERC20PaymentTerminal, mockJbFeeGauge } = await setup();
-
-    await expect(
-      jbERC20PaymentTerminal.connect(caller).setFeeGauge(mockJbFeeGauge.address),
-    ).to.be.revertedWith('Ownable: caller is not the owner');
-  });
-});
