+++ test/jb_payment_terminal/add_to_balance_of.test.js
@@ -6,35 +6,21 @@ import { makeSplits, packFundingCycleMetadata, setBalance } from '../helpers/uti
 import errors from '../helpers/errors.json';
 
 import jbDirectory from '../../artifacts/contracts/JBDirectory.sol/JBDirectory.json';
-import JbERC20PaymentTerminal from '../../artifacts/contracts/JBERC20PaymentTerminal.sol/JBERC20PaymentTerminal.json';
+import JbEthPaymentTerminal from '../../artifacts/contracts/JBETHPaymentTerminal.sol/JBETHPaymentTerminal.json';
 import jbPaymentTerminalStore from '../../artifacts/contracts/JBPaymentTerminalStore.sol/JBPaymentTerminalStore.json';
 import jbOperatoreStore from '../../artifacts/contracts/JBOperatorStore.sol/JBOperatorStore.json';
 import jbProjects from '../../artifacts/contracts/JBProjects.sol/JBProjects.json';
 import jbSplitsStore from '../../artifacts/contracts/JBSplitsStore.sol/JBSplitsStore.json';
 import jbToken from '../../artifacts/contracts/JBToken.sol/JBToken.json';
 
-
-
-
-// TODO ADD WTF_U_DOIN TEST
-
-
-
-
-
-describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
+describe('JBPaymentTerminal::addToBalanceOf(...)', function () {
   const PROJECT_ID = 2;
   const AMOUNT = ethers.utils.parseEther('10');
-  const DEFAULT_FEE = 50000000; // 5%
-
   const MIN_TOKEN_REQUESTED = 0;
   const MEMO = 'Memo Test';
 
   let CURRENCY_ETH;
-  let CURRENCY_USD;
   let ETH_PAYOUT_INDEX;
-  let MAX_FEE;
-  let AMOUNT_MINUS_FEES;
 
   before(async function () {
     let jbSplitsGroupsFactory = await ethers.getContractFactory('JBSplitsGroups');
@@ -45,12 +31,6 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
     const jbCurrenciesFactory = await ethers.getContractFactory('JBCurrencies');
     const jbCurrencies = await jbCurrenciesFactory.deploy();
     CURRENCY_ETH = await jbCurrencies.ETH();
-    CURRENCY_USD = await jbCurrencies.USD();
-
-    let jbConstantsFactory = await ethers.getContractFactory('JBConstants');
-    let jbConstants = await jbConstantsFactory.deploy();
-    MAX_FEE = (await jbConstants.MAX_FEE()).toNumber();
-    AMOUNT_MINUS_FEES = (AMOUNT.mul(MAX_FEE).div( (DEFAULT_FEE + MAX_FEE) ));
   });
 
   async function setup() {
@@ -61,27 +41,31 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
     const block = await ethers.provider.getBlock(blockNum);
     const timestamp = block.timestamp;
 
+    const SPLITS_GROUP = 1;
+
     let [
       mockJbDirectory,
-      mockJbERC20PaymentTerminal,
+      mockJbEthPaymentTerminal,
       mockJbPaymentTerminalStore,
       mockJbOperatorStore,
       mockJbProjects,
       mockJbSplitsStore,
-      mockToken
+      mockJbToken
     ] = await Promise.all([
       deployMockContract(deployer, jbDirectory.abi),
-      deployMockContract(deployer, JbERC20PaymentTerminal.abi),
+      deployMockContract(deployer, JbEthPaymentTerminal.abi),
       deployMockContract(deployer, jbPaymentTerminalStore.abi),
       deployMockContract(deployer, jbOperatoreStore.abi),
       deployMockContract(deployer, jbProjects.abi),
       deployMockContract(deployer, jbSplitsStore.abi),
-      deployMockContract(deployer, jbToken.abi),
+      deployMockContract(deployer, jbToken.abi)
     ]);
 
-    let jbTerminalFactory = await ethers.getContractFactory('JBERC20PaymentTerminal', deployer);
+    let jbTerminalFactory = await ethers.getContractFactory('JBETHPaymentTerminal', deployer);
+    let jbErc20TerminalFactory = await ethers.getContractFactory('JBERC20PaymentTerminal', deployer);
+    const NON_ETH_TOKEN = mockJbToken.address;
 
-    const currentNonce = await ethers.provider.getTransactionCount(deployer.address);
+    let currentNonce = await ethers.provider.getTransactionCount(deployer.address);
     const futureTerminalAddress = ethers.utils.getContractAddress({
       from: deployer.address,
       nonce: currentNonce + 1,
@@ -89,13 +73,34 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
 
     await mockJbPaymentTerminalStore.mock.claimFor.withArgs(futureTerminalAddress).returns();
 
-    let jbERC20PaymentTerminal = await jbTerminalFactory
+    let jbEthPaymentTerminal = await jbTerminalFactory
       .connect(deployer)
       .deploy(
-        mockToken.address,
-        /*currency*/ CURRENCY_USD, //terminal uses an usd values of mockToken, converted in eth to issue project token
-        /*base weight currency*/ CURRENCY_ETH,
-        /*payoutsplitsgroup*/ETH_PAYOUT_INDEX,
+        /*base weight currency*/CURRENCY_ETH,
+        mockJbOperatorStore.address,
+        mockJbProjects.address,
+        mockJbDirectory.address,
+        mockJbSplitsStore.address,
+        mockJbPaymentTerminalStore.address,
+        terminalOwner.address,
+      );
+
+    currentNonce = await ethers.provider.getTransactionCount(deployer.address);
+
+    const futureOtherCurrencyTerminalAddress = ethers.utils.getContractAddress({
+      from: deployer.address,
+      nonce: currentNonce + 1,
+    });
+
+    await mockJbPaymentTerminalStore.mock.claimFor.withArgs(futureOtherCurrencyTerminalAddress).returns();
+
+    let jbErc20PaymentTerminal = await jbErc20TerminalFactory
+      .connect(deployer)
+      .deploy(
+        NON_ETH_TOKEN,
+        CURRENCY_ETH,
+        CURRENCY_ETH,
+        SPLITS_GROUP,
         mockJbOperatorStore.address,
         mockJbProjects.address,
         mockJbDirectory.address,
@@ -117,7 +122,11 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
     };
 
     await mockJbDirectory.mock.isTerminalOf
-      .withArgs(PROJECT_ID, jbERC20PaymentTerminal.address)
+      .withArgs(PROJECT_ID, jbEthPaymentTerminal.address)
+      .returns(true);
+
+    await mockJbDirectory.mock.isTerminalOf
+      .withArgs(PROJECT_ID, jbErc20PaymentTerminal.address)
       .returns(true);
 
     await mockJbPaymentTerminalStore.mock.recordDistributionFor
@@ -143,7 +152,7 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
       .withArgs(PROJECT_ID, AMOUNT)
       .returns(fundingCycle);
 
-    await setBalance(jbERC20PaymentTerminal.address, AMOUNT);
+    await setBalance(jbEthPaymentTerminal.address, AMOUNT);
 
     return {
       deployer,
@@ -153,13 +162,14 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
       beneficiaryOne,
       beneficiaryTwo,
       addrs,
-      jbERC20PaymentTerminal,
+      jbEthPaymentTerminal,
+      jbErc20PaymentTerminal,
       mockJbDirectory,
-      mockJbERC20PaymentTerminal,
+      mockJbEthPaymentTerminal,
       mockJbPaymentTerminalStore,
+      mockJbToken,
       mockJbOperatorStore,
       mockJbSplitsStore,
-      mockToken,
       timestamp,
       fundingCycle,
     };
@@ -170,10 +180,9 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
       caller,
       beneficiaryOne,
       beneficiaryTwo,
-      jbERC20PaymentTerminal,
+      jbEthPaymentTerminal,
       timestamp,
       mockJbSplitsStore,
-      mockToken
     } = await setup();
     const splits = makeSplits({
       count: 2,
@@ -184,42 +193,51 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
       .withArgs(PROJECT_ID, timestamp, ETH_PAYOUT_INDEX)
       .returns(splits);
 
-    // Split distribution
-    await Promise.all(
-      splits.map(async (split) => {
-        console.log('term ', jbERC20PaymentTerminal.address);
-        console.log('spl benef', split.beneficiary);
-        console.log('net amt ',AMOUNT_MINUS_FEES.div(2).toString());
-        console.log('token :', mockToken.address);
-        await mockToken.mock.transferFrom.withArgs(jbERC20PaymentTerminal.address, split.beneficiary, AMOUNT_MINUS_FEES.div(2));
-      }
-    ))
-    
-    await jbERC20PaymentTerminal
+    await jbEthPaymentTerminal
       .connect(caller)
       .distributePayoutsOf(PROJECT_ID, AMOUNT, ETH_PAYOUT_INDEX, MIN_TOKEN_REQUESTED, MEMO);
 
-    // Initial fund sends
-    await mockToken.mock.transferFrom.withArgs(caller.address, jbERC20PaymentTerminal.address, AMOUNT);
-
-    await Promise.all(
-      splits.map(async (split) => {
-        console.log(split.beneficiary);
-        console.log(jbERC20PaymentTerminal.address);
-        console.log(AMOUNT_MINUS_FEES.div(2).toString());
-        await mockToken.mock.transferFrom.withArgs(split.beneficiary, jbERC20PaymentTerminal.address, AMOUNT_MINUS_FEES.div(2));
-      }
-    ))
-
     expect(
-      await jbERC20PaymentTerminal
+      await jbEthPaymentTerminal
         .connect(caller)
-        .addToBalanceOf(AMOUNT, PROJECT_ID, MEMO),
+        .addToBalanceOf(AMOUNT, PROJECT_ID, MEMO, { value: AMOUNT }),
     )
-      .to.emit(jbERC20PaymentTerminal, 'AddToBalance')
+      .to.emit(jbEthPaymentTerminal, 'AddToBalance')
       .withArgs(PROJECT_ID, AMOUNT, MEMO, caller.address);
 
-    expect(await jbERC20PaymentTerminal.heldFeesOf(PROJECT_ID)).to.eql([]);
+    expect(await jbEthPaymentTerminal.heldFeesOf(PROJECT_ID)).to.eql([]);
+  });
+  it('Should work with eth terminal with non msg.value amount sent', async function () {
+    const {
+      caller,
+      jbEthPaymentTerminal,
+      mockJbPaymentTerminalStore,
+      fundingCycle
+    } = await setup();
+    await mockJbPaymentTerminalStore.mock.recordAddedBalanceFor
+      .withArgs(PROJECT_ID, AMOUNT)
+      .returns(fundingCycle);
+
+    await jbEthPaymentTerminal
+      .connect(caller)
+      .addToBalanceOf(AMOUNT + 1, PROJECT_ID, MEMO, { value: AMOUNT });
+  });
+  it('Should work with non-eth terminal if no value is sent', async function () {
+    const {
+      caller,
+      jbErc20PaymentTerminal,
+      mockJbToken,
+      mockJbPaymentTerminalStore,
+      fundingCycle
+    } = await setup();
+    await mockJbPaymentTerminalStore.mock.recordAddedBalanceFor
+      .withArgs(PROJECT_ID, AMOUNT)
+      .returns(fundingCycle);
+
+    await mockJbToken.mock.transferFrom.withArgs(caller.address, jbErc20PaymentTerminal.address, AMOUNT).returns(0);
+    await jbErc20PaymentTerminal
+      .connect(caller)
+      .addToBalanceOf(AMOUNT, PROJECT_ID, MEMO, { value: 0 });
   });
 
   it('Should add to the project balance, refund a held fee by substracting the amount from the held fee amount and emit event', async function () {
@@ -227,7 +245,7 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
       caller,
       beneficiaryOne,
       beneficiaryTwo,
-      jbERC20PaymentTerminal,
+      jbEthPaymentTerminal,
       timestamp,
       mockJbSplitsStore,
       mockJbPaymentTerminalStore,
@@ -242,7 +260,7 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
       .withArgs(PROJECT_ID, timestamp, ETH_PAYOUT_INDEX)
       .returns(splits);
 
-    await jbERC20PaymentTerminal
+    await jbEthPaymentTerminal
       .connect(caller)
       .distributePayoutsOf(PROJECT_ID, AMOUNT, ETH_PAYOUT_INDEX, MIN_TOKEN_REQUESTED, MEMO);
 
@@ -250,15 +268,15 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
       .withArgs(PROJECT_ID, 1)
       .returns(fundingCycle);
 
-    let heldFeeBefore = await jbERC20PaymentTerminal.heldFeesOf(PROJECT_ID);
+    let heldFeeBefore = await jbEthPaymentTerminal.heldFeesOf(PROJECT_ID);
 
     expect(
-      await jbERC20PaymentTerminal.connect(caller).addToBalanceOf(1, PROJECT_ID, MEMO, { value: 1 }),
+      await jbEthPaymentTerminal.connect(caller).addToBalanceOf(1, PROJECT_ID, MEMO, { value: 1 }),
     )
-      .to.emit(jbERC20PaymentTerminal, 'AddToBalance')
+      .to.emit(jbEthPaymentTerminal, 'AddToBalance')
       .withArgs(PROJECT_ID, 1, MEMO, caller.address);
 
-    let heldFeeAfter = await jbERC20PaymentTerminal.heldFeesOf(PROJECT_ID);
+    let heldFeeAfter = await jbEthPaymentTerminal.heldFeesOf(PROJECT_ID);
     expect(heldFeeAfter[0].amount).to.equal(heldFeeBefore[0].amount.sub(1));
   });
 
@@ -267,7 +285,7 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
       caller,
       beneficiaryOne,
       beneficiaryTwo,
-      jbERC20PaymentTerminal,
+      jbEthPaymentTerminal,
       timestamp,
       mockJbSplitsStore,
       mockJbPaymentTerminalStore,
@@ -299,11 +317,11 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
         AMOUNT.div(2),
       );
 
-    await jbERC20PaymentTerminal
+    await jbEthPaymentTerminal
       .connect(caller)
       .distributePayoutsOf(PROJECT_ID, AMOUNT.div(2), ETH_PAYOUT_INDEX, MIN_TOKEN_REQUESTED, MEMO);
 
-    await jbERC20PaymentTerminal
+    await jbEthPaymentTerminal
       .connect(caller)
       .distributePayoutsOf(PROJECT_ID, AMOUNT.div(2), ETH_PAYOUT_INDEX, MIN_TOKEN_REQUESTED, MEMO);
 
@@ -311,28 +329,38 @@ describe('JBERC20PaymentTerminal::addToBalanceOf(...)', function () {
       .withArgs(PROJECT_ID, 10)
       .returns(fundingCycle);
 
-    let heldFeeBefore = await jbERC20PaymentTerminal.heldFeesOf(PROJECT_ID);
+    let heldFeeBefore = await jbEthPaymentTerminal.heldFeesOf(PROJECT_ID);
 
     expect(
-      await jbERC20PaymentTerminal.connect(caller).addToBalanceOf(10, PROJECT_ID, MEMO, { value: 10 }),
+      await jbEthPaymentTerminal.connect(caller).addToBalanceOf(10, PROJECT_ID, MEMO, { value: 10 }),
     )
-      .to.emit(jbERC20PaymentTerminal, 'AddToBalance')
+      .to.emit(jbEthPaymentTerminal, 'AddToBalance')
       .withArgs(PROJECT_ID, 10, MEMO, caller.address);
 
-    let heldFeeAfter = await jbERC20PaymentTerminal.heldFeesOf(PROJECT_ID);
+    let heldFeeAfter = await jbEthPaymentTerminal.heldFeesOf(PROJECT_ID);
     expect(heldFeeAfter[0].amount).to.equal(heldFeeBefore[0].amount.sub(10));
   });
+  it("Can't add with value if terminal token isn't ETH", async function () {
+    const { caller, jbErc20PaymentTerminal } = await setup();
 
+    await expect(
+      jbErc20PaymentTerminal
+        .connect(caller)
+        .addToBalanceOf(AMOUNT, PROJECT_ID, MEMO,
+          { value: 10 },
+        ),
+    ).to.be.revertedWith(errors.NO_MSG_VALUE_ALLOWED);
+  });
   it("Can't add to balance if terminal doesn't belong to project", async function () {
-    const { caller, jbERC20PaymentTerminal, mockJbDirectory } = await setup();
+    const { caller, jbEthPaymentTerminal, mockJbDirectory } = await setup();
 
     const otherProjectId = 18;
     await mockJbDirectory.mock.isTerminalOf
-      .withArgs(otherProjectId, jbERC20PaymentTerminal.address)
+      .withArgs(otherProjectId, jbEthPaymentTerminal.address)
       .returns(false);
 
     await expect(
-      jbERC20PaymentTerminal.connect(caller).addToBalanceOf(AMOUNT, otherProjectId, MEMO, { value: 0 }),
+      jbEthPaymentTerminal.connect(caller).addToBalanceOf(AMOUNT, otherProjectId, MEMO, { value: 0 }),
     ).to.be.revertedWith(errors.PROJECT_TERMINAL_MISMATCH);
   });
 });
