+++ test/jb_payment_terminal/pay.test.js
@@ -8,8 +8,9 @@ import JBPaymentTerminalStore from '../../artifacts/contracts/JBPaymentTerminalS
 import jbOperatoreStore from '../../artifacts/contracts/JBOperatorStore.sol/JBOperatorStore.json';
 import jbProjects from '../../artifacts/contracts/JBProjects.sol/JBProjects.json';
 import jbSplitsStore from '../../artifacts/contracts/JBSplitsStore.sol/JBSplitsStore.json';
+import jbToken from '../../artifacts/contracts/JBToken.sol/JBToken.json';
 
-describe('JBERC20PaymentTerminal::pay(...)', function () {
+describe('JBPaymentTerminal::pay(...)', function () {
   const PROJECT_ID = 1;
   const MEMO = 'Memo Test';
   const DELEGATE_METADATA = ethers.utils.randomBytes(32);
@@ -26,6 +27,9 @@ describe('JBERC20PaymentTerminal::pay(...)', function () {
     const block = await ethers.provider.getBlock(blockNum);
     const timestamp = block.timestamp;
 
+    const CURRENCY_ETH = 1;
+    const SPLITS_GROUP = 1;
+
     let [
       mockJbDirectory,
       mockJBPaymentTerminalStore,
@@ -40,9 +44,13 @@ describe('JBERC20PaymentTerminal::pay(...)', function () {
       deployMockContract(deployer, jbSplitsStore.abi),
     ]);
 
-    let jbTerminalFactory = await ethers.getContractFactory('JBERC20PaymentTerminal', deployer);
+    const mockJbToken = await deployMockContract(deployer, jbToken.abi);
+    const NON_ETH_TOKEN = mockJbToken.address;
 
-    const currentNonce = await ethers.provider.getTransactionCount(deployer.address);
+    let jbEthTerminalFactory = await ethers.getContractFactory('JBETHPaymentTerminal', deployer);
+    let jbErc20TerminalFactory = await ethers.getContractFactory('JBERC20PaymentTerminal', deployer);
+
+    let currentNonce = await ethers.provider.getTransactionCount(deployer.address);
     const futureTerminalAddress = ethers.utils.getContractAddress({
       from: deployer.address,
       nonce: currentNonce + 1,
@@ -50,9 +58,35 @@ describe('JBERC20PaymentTerminal::pay(...)', function () {
 
     await mockJBPaymentTerminalStore.mock.claimFor.withArgs(futureTerminalAddress).returns();
 
-    let jbERC20PaymentTerminal = await jbTerminalFactory
+    let jbEthPaymentTerminal = await jbEthTerminalFactory
       .connect(deployer)
       .deploy(
+        CURRENCY_ETH,
+        mockJbOperatorStore.address,
+        mockJbProjects.address,
+        mockJbDirectory.address,
+        mockJbSplitsStore.address,
+        mockJBPaymentTerminalStore.address,
+        terminalOwner.address,
+      );
+
+
+    currentNonce = await ethers.provider.getTransactionCount(deployer.address);
+
+    const futureOtherCurrencyTerminalAddress = ethers.utils.getContractAddress({
+      from: deployer.address,
+      nonce: currentNonce + 1,
+    });
+
+    await mockJBPaymentTerminalStore.mock.claimFor.withArgs(futureOtherCurrencyTerminalAddress).returns();
+
+    let jbErc20PaymentTerminal = await jbErc20TerminalFactory
+      .connect(deployer)
+      .deploy(
+        NON_ETH_TOKEN,
+        CURRENCY_ETH,
+        CURRENCY_ETH,
+        SPLITS_GROUP,
         mockJbOperatorStore.address,
         mockJbProjects.address,
         mockJbDirectory.address,
@@ -62,7 +96,11 @@ describe('JBERC20PaymentTerminal::pay(...)', function () {
       );
 
     await mockJbDirectory.mock.isTerminalOf
-      .withArgs(PROJECT_ID, jbERC20PaymentTerminal.address)
+      .withArgs(PROJECT_ID, jbEthPaymentTerminal.address)
+      .returns(true);
+
+    await mockJbDirectory.mock.isTerminalOf
+      .withArgs(PROJECT_ID, jbErc20PaymentTerminal.address)
       .returns(true);
 
     await mockJBPaymentTerminalStore.mock.recordPaymentFrom
@@ -99,7 +137,9 @@ describe('JBERC20PaymentTerminal::pay(...)', function () {
       caller,
       beneficiary,
       addrs,
-      jbERC20PaymentTerminal,
+      jbEthPaymentTerminal,
+      jbErc20PaymentTerminal,
+      mockJbToken,
       mockJbDirectory,
       mockJBPaymentTerminalStore,
       timestamp,
@@ -107,12 +147,13 @@ describe('JBERC20PaymentTerminal::pay(...)', function () {
   }
 
   it('Should record payment and emit event', async function () {
-    const { caller, jbERC20PaymentTerminal, timestamp } = await setup();
+    const { caller, jbEthPaymentTerminal, timestamp } = await setup();
 
     expect(
-      await jbERC20PaymentTerminal
+      await jbEthPaymentTerminal
         .connect(caller)
         .pay(
+          ETH_TO_PAY,
           PROJECT_ID,
           caller.address,
           MIN_TOKEN_REQUESTED,
@@ -122,7 +163,7 @@ describe('JBERC20PaymentTerminal::pay(...)', function () {
           { value: ETH_TO_PAY },
         ),
     )
-      .to.emit(jbERC20PaymentTerminal, 'Pay')
+      .to.emit(jbEthPaymentTerminal, 'Pay')
       .withArgs(
         /*fundingCycle.configuration=*/ timestamp,
         FUNDING_CYCLE_NUMBER,
@@ -136,13 +177,68 @@ describe('JBERC20PaymentTerminal::pay(...)', function () {
       );
   });
 
-  it("Can't send tokens to the zero address", async function () {
-    const { caller, jbERC20PaymentTerminal } = await setup();
+  it('Should work with eth terminal with non msg.value amount sent', async function () {
+    const { caller, jbEthPaymentTerminal, timestamp } = await setup();
+
+    await jbEthPaymentTerminal
+      .connect(caller)
+      .pay(
+        ETH_TO_PAY + 1,
+        PROJECT_ID,
+        caller.address,
+        MIN_TOKEN_REQUESTED,
+          /*preferClaimedToken=*/ true,
+        MEMO,
+        DELEGATE_METADATA,
+        { value: ETH_TO_PAY },
+      );
+  });
+
+  it('Should work with non-eth terminal if no value is sent', async function () {
+    const { caller, jbErc20PaymentTerminal, mockJbToken } = await setup();
+
+    await mockJbToken.mock.transferFrom.withArgs(caller.address, jbErc20PaymentTerminal.address, ETH_TO_PAY).returns(0);
+    await jbErc20PaymentTerminal
+      .connect(caller)
+      .pay(
+        ETH_TO_PAY,
+        PROJECT_ID,
+        caller.address,
+        MIN_TOKEN_REQUESTED,
+          /*preferClaimedToken=*/ true,
+        MEMO,
+        DELEGATE_METADATA,
+        { value: 0 },
+      );
+  });
+
+  it("Can't pay with value if terminal token isn't ETH", async function () {
+    const { caller, jbErc20PaymentTerminal } = await setup();
 
     await expect(
-      jbERC20PaymentTerminal
+      jbErc20PaymentTerminal
         .connect(caller)
         .pay(
+          ETH_TO_PAY,
+          PROJECT_ID,
+          ethers.constants.AddressZero,
+          MIN_TOKEN_REQUESTED,
+          /*preferClaimedToken=*/ true,
+          MEMO,
+          DELEGATE_METADATA,
+          { value: ETH_TO_PAY },
+        ),
+    ).to.be.revertedWith(errors.NO_MSG_VALUE_ALLOWED);
+  });
+
+  it("Can't send tokens to the zero address", async function () {
+    const { caller, jbEthPaymentTerminal } = await setup();
+
+    await expect(
+      jbEthPaymentTerminal
+        .connect(caller)
+        .pay(
+          ETH_TO_PAY,
           PROJECT_ID,
           ethers.constants.AddressZero,
           MIN_TOKEN_REQUESTED,
@@ -155,17 +251,18 @@ describe('JBERC20PaymentTerminal::pay(...)', function () {
   });
 
   it("Can't pay if current terminal doesn't belong to project", async function () {
-    const { caller, jbERC20PaymentTerminal, mockJbDirectory } = await setup();
+    const { caller, jbEthPaymentTerminal, mockJbDirectory } = await setup();
 
     const otherProjectId = 18;
     await mockJbDirectory.mock.isTerminalOf
-      .withArgs(otherProjectId, jbERC20PaymentTerminal.address)
+      .withArgs(otherProjectId, jbEthPaymentTerminal.address)
       .returns(false);
 
     await expect(
-      jbERC20PaymentTerminal
+      jbEthPaymentTerminal
         .connect(caller)
         .pay(
+          ETH_TO_PAY,
           otherProjectId,
           ethers.constants.AddressZero,
           MIN_TOKEN_REQUESTED,
