+++ test/jb_payment_terminal/toggle_feeless_terminal.test.js
@@ -4,33 +4,37 @@ import { ethers } from 'hardhat';
 import { deployMockContract } from '@ethereum-waffle/mock-contract';
 
 import jbDirectory from '../../artifacts/contracts/interfaces/IJBDirectory.sol/IJBDirectory.json';
-import JbERC20PaymentTerminal from '../../artifacts/contracts/JBERC20PaymentTerminal.sol/JBERC20PaymentTerminal.json';
+import JbEthPaymentTerminal from '../../artifacts/contracts/JBETHPaymentTerminal.sol/JBETHPaymentTerminal.json';
 import JBPaymentTerminalStore from '../../artifacts/contracts/JBPaymentTerminalStore.sol/JBPaymentTerminalStore.json';
 import jbOperatoreStore from '../../artifacts/contracts/interfaces/IJBOperatorStore.sol/IJBOperatorStore.json';
 import jbProjects from '../../artifacts/contracts/interfaces/IJBProjects.sol/IJBProjects.json';
 import jbSplitsStore from '../../artifacts/contracts/interfaces/IJBSplitsStore.sol/IJBSplitsStore.json';
 
-describe('JBERC20PaymentTerminal::toggleFeelessTerminal(...)', function () {
+describe('JBPaymentTerminal::toggleFeelessTerminal(...)', function () {
   async function setup() {
     let [deployer, terminalOwner, caller] = await ethers.getSigners();
 
     let [
       mockJbDirectory,
-      mockJbERC20PaymentTerminal,
+      mockJbEthPaymentTerminal,
       mockJBPaymentTerminalStore,
       mockJbOperatorStore,
       mockJbProjects,
       mockJbSplitsStore,
     ] = await Promise.all([
       deployMockContract(deployer, jbDirectory.abi),
-      deployMockContract(deployer, JbERC20PaymentTerminal.abi),
+      deployMockContract(deployer, JbEthPaymentTerminal.abi),
       deployMockContract(deployer, JBPaymentTerminalStore.abi),
       deployMockContract(deployer, jbOperatoreStore.abi),
       deployMockContract(deployer, jbProjects.abi),
       deployMockContract(deployer, jbSplitsStore.abi),
     ]);
 
-    let jbTerminalFactory = await ethers.getContractFactory('JBERC20PaymentTerminal', deployer);
+    const jbCurrenciesFactory = await ethers.getContractFactory('JBCurrencies');
+    const jbCurrencies = await jbCurrenciesFactory.deploy();
+    const CURRENCY_ETH = await jbCurrencies.ETH();
+
+    let jbTerminalFactory = await ethers.getContractFactory('JBETHPaymentTerminal', deployer);
 
     const currentNonce = await ethers.provider.getTransactionCount(deployer.address);
     const futureTerminalAddress = ethers.utils.getContractAddress({
@@ -40,9 +44,10 @@ describe('JBERC20PaymentTerminal::toggleFeelessTerminal(...)', function () {
 
     await mockJBPaymentTerminalStore.mock.claimFor.withArgs(futureTerminalAddress).returns();
 
-    let jbERC20PaymentTerminal = await jbTerminalFactory
+    let jbEthPaymentTerminal = await jbTerminalFactory
       .connect(deployer)
       .deploy(
+        CURRENCY_ETH,
         mockJbOperatorStore.address,
         mockJbProjects.address,
         mockJbDirectory.address,
@@ -54,30 +59,30 @@ describe('JBERC20PaymentTerminal::toggleFeelessTerminal(...)', function () {
     return {
       terminalOwner,
       caller,
-      jbERC20PaymentTerminal,
-      mockJbERC20PaymentTerminal,
+      jbEthPaymentTerminal,
+      mockJbEthPaymentTerminal,
     };
   }
 
   it('Should add a terminal as feeless and emit event, if the terminal was not feeless before', async function () {
-    const { terminalOwner, jbERC20PaymentTerminal, mockJbERC20PaymentTerminal } = await setup();
+    const { terminalOwner, jbEthPaymentTerminal, mockJbEthPaymentTerminal } = await setup();
 
-    expect(await jbERC20PaymentTerminal.connect(terminalOwner).toggleFeelessTerminal(mockJbERC20PaymentTerminal.address))
-      .to.emit(jbERC20PaymentTerminal, 'SetFeelessTerminal')
-      .withArgs(mockJbERC20PaymentTerminal.address, terminalOwner.address);
-    
-    expect(await jbERC20PaymentTerminal.isFeelessTerminal(mockJbERC20PaymentTerminal.address)).to.be.true;
+    expect(await jbEthPaymentTerminal.connect(terminalOwner).toggleFeelessTerminal(mockJbEthPaymentTerminal.address))
+      .to.emit(jbEthPaymentTerminal, 'SetFeelessTerminal')
+      .withArgs(mockJbEthPaymentTerminal.address, terminalOwner.address);
+
+    expect(await jbEthPaymentTerminal.isFeelessTerminal(mockJbEthPaymentTerminal.address)).to.be.true;
   });
 
   it('Should remove a terminal as feeless and emit event, if the terminal was feeless before', async function () {
-    const { terminalOwner, jbERC20PaymentTerminal, mockJbERC20PaymentTerminal } = await setup();
+    const { terminalOwner, jbEthPaymentTerminal, mockJbEthPaymentTerminal } = await setup();
 
-    await jbERC20PaymentTerminal.connect(terminalOwner).toggleFeelessTerminal(mockJbERC20PaymentTerminal.address);
+    await jbEthPaymentTerminal.connect(terminalOwner).toggleFeelessTerminal(mockJbEthPaymentTerminal.address);
 
-    expect(await jbERC20PaymentTerminal.connect(terminalOwner).toggleFeelessTerminal(mockJbERC20PaymentTerminal.address))
-      .to.emit(jbERC20PaymentTerminal, 'SetFeelessTerminal')
-      .withArgs(mockJbERC20PaymentTerminal.address, terminalOwner.address);
-    
-    expect(await jbERC20PaymentTerminal.isFeelessTerminal(mockJbERC20PaymentTerminal.address)).to.be.false;
+    expect(await jbEthPaymentTerminal.connect(terminalOwner).toggleFeelessTerminal(mockJbEthPaymentTerminal.address))
+      .to.emit(jbEthPaymentTerminal, 'SetFeelessTerminal')
+      .withArgs(mockJbEthPaymentTerminal.address, terminalOwner.address);
+
+    expect(await jbEthPaymentTerminal.isFeelessTerminal(mockJbEthPaymentTerminal.address)).to.be.false;
   });
 });
